appium-python/requirements.txt
Appium-Python-Client==0.22
py==1.4.31
pytest==2.9.2
selenium==2.53.5
